Fix https://github.com/emoncms/emoncms/issues/138
Basically `#` characters are now automatically removed in the designer, and re-added within the plot function. This fixes the issue where having a `#` character in the URL (which is how plot parameters are passed to each plot iframe) causes the url to be invalid, since `#` has special meaning in a URL.

Lastly, indentation was converted from 2 to 4 spaces.