Fix saving when the content hasn't actually changed
See #31. There was a problem where if you saved when the content hadn't actually changed, then there was an error. Of course it's not actually an error, but rather the content hasn't changed, so there's nothing to save. This could happen if you moved something to the left and then back again. With undo/redo now it's more likely to happen, so important to fix. This fixes it.