Optimized view for negative values
Additionally to the request "Optimized value below the bar" we should not forget negative values for setup the punctiation.
Using the graduation there is a similar behavior. I prepared a solution to view the complete text. The only weak point is that the scaling could have different formats like -100°C and 100.0°C in one object.

TOM